Move message handlers to their own classes so that they can be instantiated.
