+++ src/main/java/kihira/playerbeacons/proxy/CommonProxy.java
@@ -4,6 +4,7 @@ import cpw.mods.fml.relauncher.Side;
 import kihira.playerbeacons.api.buff.Buff;
 import kihira.playerbeacons.common.PlayerBeacons;
 import kihira.playerbeacons.common.network.CorruptionUpdateMessage;
+import kihira.playerbeacons.common.network.CorruptionUpdateMessageHandler;
 import kihira.playerbeacons.common.network.PlayerPhasedMessage;
 import kihira.playerbeacons.common.network.PlayerPhasedMessageHandler;
 import kihira.playerbeacons.common.tileentity.TileEntityPlayerBeacon;
@@ -12,7 +13,7 @@ import net.minecraft.client.entity.AbstractClientPlayer;
 public class CommonProxy {
 
     public void registerMessages() {
-        PlayerBeacons.networkWrapper.registerMessage(CorruptionUpdateMessage.CorruptionUpdateMessageHandler.class, CorruptionUpdateMessage.class, 0, Side.SERVER);
+        PlayerBeacons.networkWrapper.registerMessage(CorruptionUpdateMessageHandler.class, CorruptionUpdateMessage.class, 0, Side.SERVER);
         PlayerBeacons.networkWrapper.registerMessage(PlayerPhasedMessageHandler.class, PlayerPhasedMessage.class, 1, Side.SERVER);
     }
